[REFACTOR] #251: 피드백 반영
